backend/pyproject.toml
[project]
name = "dept-website-backend"
version = "0.1.0"
description = "FastAPI backend for AI-enhanced department website"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
    "uvicorn>=0.32.0",
    "pyjwt>=2.9.0",
    "python-multipart>=0.0.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
